refactor: split into component
