+++ client/src/pages/Signup.js
@@ -1,67 +1,10 @@
-import React, {Fragment, useState} from "react";
-import {TextField, Button, OutlinedInput, FormHelperText, Typography, Box, Grid} from "@material-ui/core";
+import React, {useState} from "react";
+import {Grid} from "@material-ui/core";
 import { makeStyles } from '@material-ui/core/styles';
-import {Link, useHistory} from "react-router-dom";
-import { useForm } from 'react-hook-form';
+import {useHistory} from "react-router-dom";
+import { AuthForm, RedirectDiv} from "../Components/auth"
 import {authStyle} from '../themes/signup.style';
 import axios from 'axios';
-
-const AuthForm = (props) => {
-    const { handleSubmit, register, errors } = useForm();
-
-    const emailValidator = {
-        required: 'Email is required',
-        pattern: {
-            value: /^\S+@\S+$/i,
-            message: "Invalid email address"
-        }
-    }
-
-    const passwordValidator = {
-        required: 'Password is required',
-        minLength: {
-            value: 6,
-            message: "Password needs 6 characters at least"
-        }
-    }
-
-    const classes = makeStyles(authStyle.form)();
-
-    return (
-        <Box component={'form'} className={classes.root} onSubmit={handleSubmit(props.onSubmit)}>
-            <Typography align={'center'} className={classes.title}>{props.title}</Typography>
-
-            <FormHelperText error>{props.serverResponse}</FormHelperText>
-
-            <TextField name="email" variant={'outlined'} type="text" label={props.input1}
-                       className={classes.input}
-                       inputRef={register(emailValidator)}
-                       error={errors.email}
-                       helperText={errors.email && errors.email.message}
-            />
-
-            <TextField name="password" variant={'outlined'} type="password" label={props.input2}
-                       className={classes.input}
-                       inputRef={register(passwordValidator)}
-                       error={errors.password}
-                       helperText={errors.password && errors.password.message}
-            />
-
-            <Button color={'primary'} className={classes.button} variant={'contained'} type={'submit'}> {props.submit}</Button>
-        </Box>
-    )
-};
-
-const RedirectDiv = (props) => {
-    const AuthDivStyle = makeStyles(authStyle.div)();
-
-    return (
-        <Box component={'div'} className={AuthDivStyle.root}>
-            <Typography className={AuthDivStyle.margin}>{props.title}</Typography>
-            <Link to={props.link}><Typography>{props.desc}</Typography></Link>
-        </Box>
-    )
-}
 
 const Signup = () => {
     const history = useHistory();
@@ -80,8 +23,9 @@ const Signup = () => {
         const url = '/api/v1/auth/register'
         axios.post(url, {email, password})
             .then(res => {
+
                 //If success to create a new account, redirect to login page
-                if (res.data.success){
+                if (res.status === 200){
                     history.push('/login')
                 } else {
                     setServerResponse(res.data.error);
@@ -97,9 +41,9 @@ const Signup = () => {
     return (
         <Grid container className={classes.vh100}>
             <Grid item xs={6} >
-                <img src={"/images/image1.png"} width={'100%'} height={'100%'}/>
+                <img alt='' src={"/images/image1.png"} width={'100%'} height={'100%'}/>
             </Grid>
-            <Grid item xs={6} alignItems={'center'}>
+            <Grid item xs={6} >
                 <AuthForm title="Sign up to Kanban" input1="Enter email" input2="Create Password" submit="Sign up"
                           onSubmit={onSubmit}
                           serverResponse={serverResponse}/>
@@ -109,8 +53,4 @@ const Signup = () => {
     )
 }
 
-export {
-    Signup,
-    AuthForm,
-    RedirectDiv
-};
+export {Signup};
